+++ tests/Stream/OutputStreamTest.php
@@ -2,24 +2,25 @@
 declare(strict_types=1);
 
 /**
- * Lupin package.
+ * GpsLab component.
  *
  * @author    Peter Gribanov <info@peter-gribanov.ru>
- * @copyright Copyright (c) 2011, Peter Gribanov
+ * @copyright Copyright (c) 2011-2019, Peter Gribanov
+ * @license   http://opensource.org/licenses/MIT
  */
 
-namespace GpsLab\Component\Sitemap\Tests\Unit\Stream;
+namespace GpsLab\Component\Sitemap\Tests\Stream;
 
 use GpsLab\Component\Sitemap\Render\SitemapRender;
-use GpsLab\Component\Sitemap\Stream\CallbackStream;
 use GpsLab\Component\Sitemap\Stream\Exception\LinksOverflowException;
 use GpsLab\Component\Sitemap\Stream\Exception\SizeOverflowException;
 use GpsLab\Component\Sitemap\Stream\Exception\StreamStateException;
+use GpsLab\Component\Sitemap\Stream\OutputStream;
 use GpsLab\Component\Sitemap\Url\Url;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-class CallbackStreamTest extends TestCase
+class OutputStreamTest extends TestCase
 {
     /**
      * @var MockObject|SitemapRender
@@ -27,7 +28,7 @@ class CallbackStreamTest extends TestCase
     private $render;
 
     /**
-     * @var CallbackStream
+     * @var OutputStream
      */
     private $stream;
 
@@ -41,18 +42,24 @@ class CallbackStreamTest extends TestCase
      */
     private const CLOSED = 'Stream closed';
 
+    /**
+     * @var string
+     */
+    private $expected_buffer = '';
+
     protected function setUp(): void
     {
         $this->render = $this->createMock(SitemapRender::class);
-        $call = 0;
-        $this->stream = new CallbackStream($this->render, function ($content) use (&$call) {
-            if ($call === 0) {
-                self::assertEquals(self::OPENED, $content);
-            } else {
-                self::assertEquals(self::CLOSED, $content);
-            }
-            ++$call;
-        });
+
+        $this->stream = new OutputStream($this->render);
+        ob_start();
+    }
+
+    protected function tearDown(): void
+    {
+        self::assertEquals($this->expected_buffer, ob_get_clean());
+        $this->expected_buffer = '';
+        ob_clean();
     }
 
     public function testOpenClose(): void
@@ -110,19 +117,13 @@ class CallbackStreamTest extends TestCase
 
     public function testPush(): void
     {
+        $this->open();
+
         $urls = [
             new Url('/foo'),
             new Url('/bar'),
             new Url('/baz'),
         ];
-        $call = 0;
-        $this->stream = new CallbackStream($this->render, function ($content) use (&$call, $urls) {
-            if (isset($urls[$call - 1])) {
-                self::assertEquals($urls[$call - 1]->getLoc(), $content);
-            }
-            ++$call;
-        });
-        $this->open();
 
         foreach ($urls as $i => $url) {
             /* @var $url Url */
@@ -132,6 +133,7 @@ class CallbackStreamTest extends TestCase
                 ->with($urls[$i])
                 ->will(self::returnValue($url->getLoc()))
             ;
+            $this->expected_buffer .= $url->getLoc();
         }
 
         foreach ($urls as $url) {
@@ -144,18 +146,7 @@ class CallbackStreamTest extends TestCase
     public function testOverflowLinks(): void
     {
         $loc = '/';
-        $call = 0;
-        $this->stream = new CallbackStream($this->render, function ($content) use (&$call, $loc) {
-            if ($call === 0) {
-                self::assertEquals(self::OPENED, $content);
-            } elseif ($call - 1 < CallbackStream::LINKS_LIMIT) {
-                self::assertEquals($loc, $content);
-            } else {
-                self::assertEquals(self::CLOSED, $content);
-            }
-            ++$call;
-        });
-        $this->open();
+        $this->stream->open();
         $this->render
             ->expects(self::atLeastOnce())
             ->method('url')
@@ -163,56 +154,45 @@ class CallbackStreamTest extends TestCase
         ;
 
         try {
-            for ($i = 0; $i <= CallbackStream::LINKS_LIMIT; ++$i) {
+            for ($i = 0; $i <= OutputStream::LINKS_LIMIT; ++$i) {
                 $this->stream->push(new Url($loc));
             }
             self::assertTrue(false, 'Must throw LinksOverflowException.');
         } catch (LinksOverflowException $e) {
-            $this->close();
+            $this->stream->close();
+            ob_clean(); // not check content
         }
     }
 
     public function testOverflowSize(): void
     {
-        $i = 0;
         $loops = 10000;
-        $loop_size = (int) floor(CallbackStream::BYTE_LIMIT / $loops);
-        $prefix_size = CallbackStream::BYTE_LIMIT - ($loops * $loop_size);
-        $opened = str_repeat('/', ++$prefix_size); // overflow byte
+        $loop_size = (int) floor(OutputStream::BYTE_LIMIT / $loops);
+        $prefix_size = OutputStream::BYTE_LIMIT - ($loops * $loop_size);
+        ++$prefix_size; // overflow byte
         $loc = str_repeat('/', $loop_size);
 
         $this->render
             ->expects(self::at(0))
             ->method('start')
-            ->will(self::returnValue($opened))
+            ->will(self::returnValue(str_repeat('/', $prefix_size)))
         ;
         $this->render
             ->expects(self::atLeastOnce())
             ->method('url')
             ->will(self::returnValue($loc))
         ;
-        $call = 0;
-        $this->stream = new CallbackStream(
-            $this->render,
-            function ($content) use (&$call, $loc, &$i, $loops, $opened) {
-                if ($call === 0) {
-                    self::assertEquals($opened, $content);
-                } elseif ($i + 1 < $loops) {
-                    self::assertEquals($loc, $content);
-                }
-                ++$call;
-            }
-        );
 
         $this->stream->open();
 
         try {
-            for (; $i < $loops; ++$i) {
+            for ($i = 0; $i < $loops; ++$i) {
                 $this->stream->push(new Url($loc));
             }
             self::assertTrue(false, 'Must throw SizeOverflowException.');
         } catch (SizeOverflowException $e) {
             $this->stream->close();
+            ob_clean(); // not check content
         }
     }
 
@@ -230,10 +210,12 @@ class CallbackStreamTest extends TestCase
         ;
 
         $this->stream->open();
+        $this->expected_buffer .= self::OPENED;
     }
 
     private function close(): void
     {
         $this->stream->close();
+        $this->expected_buffer .= self::CLOSED;
     }
 }
